Enable bumping semantic version elements. Retains default behavior of bumping the patch version with module:bump (even though the docs said it bumped the minor version). I think this would technically be a breaking change as it enforces Semantic Versioning wheras we were using Gem::Version previously, which isn't SemVer. I opted to use the vendored implementation of the 'semantic' gem in 'puppet/vendor' but I'm open to changing that.
Because of the way Puppet has decided to vendor the 'Semantic' gem, we're unable to use that namespace because it conflicts with the vendored implementation. This commit vendors our own implementation of the 'Semantic' gem in the Blacksmith::VersionHelper namespace. Not sure if this is the best way to do it, but I can't think of a better way. Need to make sure the original author is credited appropriately.