(maint) Remove outdated variable reference
The `upload` method in the `forge.rb` class was refactored, but one
error message that relied on variables which used to be defined in the
method was not updated with the new variables that are in scope for the
method. This removes the variables that now refer to nothing (and thus
throw and error when they're referenced), and replaces them with
variables that are in scope for the `upload` function.